discrete distribution

+++ src/main/java/ch/alpine/tensor/pdf/d/BorelTannerDistribution.java
@@ -1,6 +1,8 @@
 // code by jph
 package ch.alpine.tensor.pdf.d;
 
+import java.math.BigInteger;
+
 import ch.alpine.tensor.RealScalar;
 import ch.alpine.tensor.Scalar;
 import ch.alpine.tensor.Scalars;
@@ -38,19 +40,20 @@ public class BorelTannerDistribution extends EvaluatedDiscreteDistribution {
   private final Scalar n;
   private final Scalar o_alpha;
   private final Scalar logAlpha;
-  private final int _n;
+  private final BigInteger _n;
 
   private BorelTannerDistribution(Scalar alpha, int n) {
     this.alpha = alpha;
     this.n = RealScalar.of(n);
     o_alpha = RealScalar.ONE.subtract(alpha);
     logAlpha = Log.FUNCTION.apply(alpha);
-    _n = n;
+    
+    _n = BigInteger.valueOf(n);
     build(Tolerance.CHOP);
   }
 
   @Override // from DiscreteDistribution
-  public int lowerBound() {
+  public BigInteger lowerBound() {
     return _n;
   }
